moved `framework_base.py` to `framework/base.py`
