+++ pyrigi/framework/base.py
